refactor: move model tool helpers to builtin
